app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib==1.7.4
bcrypt==4.0.1
python-jose
